test(e2e): scope past-cutoff test to held-back (regime 2); drop recovery job
CI run 27482200429 showed the full recovery (regime 3) is not E2E-testable on
this stack:
- The only short-expiry option, block-based delays (<512), trips an arkd rc.1
  nil-pointer panic in Settings.Digest.
- Even past that, arkd REJECTS re-enrolling the expired coin into a recovery
  batch: INVALID_VTXO_SCRIPT "... is a deprecated key" — the coin cannot join a
  batch while its script still carries the deprecated key. Live recovery needs
  an arkd fix and/or an SDK unroll-to-onchain + recovery-script step.

Scope the past-cutoff E2E to the reliable regime-2 (held-back) half, which needs
no short expiry:
- Replace PastCutoffRecoveryRotationTests with PastCutoffHeldBackRotationTests
  (tagged RealRotation): real rotate-signer --cutoff -60, assert the coin is
  excluded from GetAvailableCoins (held back).
- Remove the e2e-rotation-recovery job + its block-delay config; publish gated on
  [build, e2e, e2e-rotation]. Recovery logic stays covered by unit tests.
