+++ .github/workflows/build.yml
@@ -46,11 +46,11 @@ jobs:
         timeout-minutes: 20
       - name: Restore and build
         run: dotnet build
-      # The live-rotation tests run in the isolated e2e-rotation[-recovery] jobs below: a real
-      # rotation recreates arkd-wallet + restarts arkd, cascade-failing every test sharing this
-      # stack. Run everything EXCEPT those (tagged RealRotation / RealRotationExpiry) here.
+      # The live-rotation tests run in the isolated e2e-rotation job below: a real rotation recreates
+      # arkd-wallet + restarts arkd, cascade-failing every test sharing this stack. Run everything
+      # EXCEPT those (tagged RealRotation) here.
       - name: Run E2E tests
-        run: dotnet test NArk.Tests.End2End --no-build --verbosity normal --filter "TestCategory!=RealRotation&TestCategory!=RealRotationExpiry"
+        run: dotnet test NArk.Tests.End2End --no-build --verbosity normal --filter "TestCategory!=RealRotation"
         timeout-minutes: 15
       - name: Collect docker logs on failure
         if: failure()
@@ -120,66 +120,10 @@ jobs:
         if: always()
         run: node regtest/regtest.mjs clean || true
 
-  e2e-rotation-recovery:
-    name: E2E Signer Rotation Recovery Test
-    runs-on: ubuntu-latest
-    needs: build
-    steps:
-      - uses: actions/checkout@v4
-        with:
-          submodules: true
-          fetch-depth: 0
-      - name: Setup .NET
-        uses: actions/setup-dotnet@v4
-        with:
-          dotnet-version: 9.0.x
-      - name: Setup Node
-        uses: actions/setup-node@v4
-        with:
-          node-version: '20'
-      # Unpin the operator signer (see e2e-rotation) AND switch arkd to short BLOCK-based delays so the
-      # past-cutoff recovery test can mine the coin to expiry within the run (the committed 1h time-based
-      # expiry would never elapse). arkd requires ALL delays to be the same TYPE — block (<512) or time
-      # (>=512) — so the WHOLE set moves below 512, mirroring the default ordering (BOARDING > VTXO =
-      # CHECKPOINT > UNILATERAL = PUBLIC). Height-based expiry = deterministic via MineBlocks, no wall-clock.
-      - name: Unpin signer + switch arkd to short block-based delays for the recovery test
-        run: |
-          sed -i -E '/^ARKD_WALLET_(SIGNER_KEY|DEPRECATED_SIGNER_KEYS)=/d' .env.regtest
-          {
-            echo "ARKD_VTXO_TREE_EXPIRY=40"
-            echo "ARKD_UNILATERAL_EXIT_DELAY=20"
-            echo "ARKD_PUBLIC_UNILATERAL_EXIT_DELAY=20"
-            echo "ARKD_BOARDING_EXIT_DELAY=80"
-            echo "ARKD_CHECKPOINT_EXIT_DELAY=40"
-          } >> .env.regtest
-      - name: Start regtest infrastructure
-        run: node regtest/regtest.mjs start --profile boltz,delegate
-        timeout-minutes: 20
-      - name: Restore and build
-        run: dotnet build
-      - name: Run past-cutoff recovery E2E test
-        run: dotnet test NArk.Tests.End2End --no-build --verbosity normal --filter "TestCategory=RealRotationExpiry"
-        timeout-minutes: 15
-      - name: Collect docker logs on failure
-        if: failure()
-        run: |
-          docker logs arkd > /tmp/arkd-logs.txt 2>&1 || true
-          docker logs arkd-wallet > /tmp/arkd-wallet-logs.txt 2>&1 || true
-          docker logs bitcoin > /tmp/bitcoin-logs.txt 2>&1 || true
-      - name: Upload logs
-        if: failure()
-        uses: actions/upload-artifact@v4
-        with:
-          name: docker-logs-rotation-recovery
-          path: /tmp/*-logs.txt
-      - name: Stop infrastructure
-        if: always()
-        run: node regtest/regtest.mjs clean || true
-
   publish:
     name: Publish NuGet Packages
     runs-on: ubuntu-latest
-    needs: [build, e2e, e2e-rotation, e2e-rotation-recovery]
+    needs: [build, e2e, e2e-rotation]
     if: github.ref == 'refs/heads/master' || startsWith(github.ref, 'refs/tags/')
     permissions:
       contents: write
